Back from edit courier is now working.
I need to unable courier editing. And i need to create a courier for adding and deleting a courier from a region.
I need to add a option to the region screen to see all couriers that are assigned to this region, I shuold to this as a link exactly has i did with the deliveries of a region. Then i need to add the option to add a courier to this region. I need to show all the availables courier that are not assigned to this region.